gros commit

+++ Shared/src/main/java/shared/messages/notifications/ConfirmationValiditeStage.java
@@ -10,8 +10,7 @@ package shared.messages.notifications;
  * @author Mahdi
  */
 public class ConfirmationValiditeStage extends Notification {
-
-    public ConfirmationValiditeStage(int idDemandeConvention) {
+    public ConfirmationValiditeStage(Long idDemandeConvention) {
         super(idDemandeConvention);
     }
